requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.2
django-appconf==1.0.5
django-compressor==3.1
django-libsass==0.9
docopt==0.6.2
idna==3.3
libsass==0.21.0
Pillow==9.0.1
pipreqs==0.4.11
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
yarg==0.1.9

djangorestframework~=3.13.1